JWT auth

+++ backend/db/repository/jobs.py
@@ -1,8 +1,9 @@
 from sqlalchemy.orm import Session
+from sqlalchemy.sql.functions import current_user
 
 from schemas.jobs import JobCreate
 from db.models.jobs import Job
-
+from db.models.users import User
 
 def create_new_job(job: JobCreate, db: Session, owner_id: int):
     job = Job(**job.dict(), owner_id=owner_id)
@@ -33,11 +34,15 @@ def update_job_by_id(id: int, job: JobCreate, db: Session, owner_id: int):
     return 1
 
 
-def delete_job_by_id(id: int, db: Session):
+def delete_job_by_id(id: int, db: Session, current_user: User):
     existing_job = db.query(Job).filter(Job.id == id)
+
     if not existing_job.first():
         return 0
     
+    if existing_job.owner_id != current_user.id and not current_user.is_superuser:
+        return 0
+
     existing_job.delete()
     db.commit()
     return 1
